Refactor: Rework PackageManager implementation (#243)
Summary:
---------

Removed nasty `class`, and replaced it with grouped methods.

`projectDir` is always the same in `cli` lifecycle, so we set it once on the start. Based on the presence of `yarn.lock` we decide either to use yarn or not to use it. `preferYarn` is overriding this behavior, and can later be used in `init` command which lacks correct `projectDir`.

Test Plan:
----------

- [x] - Unit tests passing
- [x] - Manual testing