fix: change how relative path to podfile is created
Previous implementation made assumption that path to podspec contains
path to project root, which is not always the case, for example in
monorepos. This caused incorrect generation of podspec path, since in
that case it would return absolute podspec path prepended with root path.
Used `pathname` class method `relative_path_from` to avoid this issue.

Resolves #537