[Docs][Autolinking] Fix monorepo examples relative path
The example folder structure as well as the edited paths to `node_modules` reflect a path that is only one extra level up. This PR updates the relative paths passed into the scripts to match. I noticed they don't match while I was upgrading to RN 0.60 because I have a comparable folder structure, but had issues on both iOS and Android when I initially tried passing in `../..` as described in the docs.