feat: collapse YellowBox/DevTools frames in Metro config (#780)
* Collapse YellowBox/DevTools frames in Metro config

React Native master (note: commit not pushed to GitHub yet) is no longer stripping `YellowBox` and React DevTools stack frames from warnings with the `framesToPop` mechanism, which we will be deprecating and ultimately removing from the default error reporter too. The recommended approach going forward is to rely on Metro's `customizeFrame` for all postprocessing of stack traces.

This commit updates the default Metro config to skip stack frames in `YellowBox` and React DevTools code, thus preserving the old behaviour. As for compatibility and versioning: It is harmless to run this change against older RN versions, but the next RN release should include this CLI change. (cc @cpojer)

* Fix typo in regex