fix: disable caching of assets for the development server (#1307)
It looks like the built assets get cached sometimes causing issues such as https://github.com/facebook/react-native/issues/28844
This might happen if the HTML file got cached, but there's a new version available with different links.

It's better to disable the caching altogether since the development server is run locally and caching files causes more problems than it helps.
Normally it's possible by developers to disable caching when devtools is open, but it's not reasonable to expect everyone to do it.

The commit changes 2 things:

- Set relevant headers to prevent caching of assets using the 'nocache' middleware
- When building files with parcel, make sure that the name of the built files don't change

I have tested this in a fresh React Native project following the instructions in CONTRIBUTING.md and observed that the debugger works as expected.