Fix metro crash when Flipper connects
Fixes per work around: https://github.com/facebook/flipper/issues/3189#issuecomment-997191720. The same fix has been applied to FB infra, and the next release will temporarily set an `origin` header as work around for existing RN versions. 

Fixes:
* https://github.com/facebook/flipper/issues/2870
* https://github.com/facebook/flipper/issues/3189

The essence of the fix is that we shouldn't check the `origin` header if it is not present. That might sound unsafe, but it actually is just a check to protect from malicious websites, and browsers ensure the origin header is set for cross site requests. For websocket connections created outside a browser context, the creator has full control over the origin header anyway, so checking it would not add any safety.