Fix Web Debugging and Other Websocket Endpoints with Metro 0.67 (#1560)
* Fix Web Debugging and Other Websocket Endpoints with Metro 0.67

Metro 0.67 included https://github.com/facebook/metro/commit/38a200e748c2e22eb34ce628d51bbdf563dbe777 which changed the way external applications should interface with it to add new WebSocket endpoints. This materializes as failures when web debugging RN 0.68 apps.

This change applies the reccomended fix, of passing a list of non-server endpoints to Metro when starting the server. I verified that web debugging works correctly after making the fix, for the same app where it previously did not work.

`ws` is updated as part of the change, along with removing tpe-checker suppressions, of which one was resulting in a runtime error.

* Remove session number

* chore: stop using stubs for isDebuggerConnected and broadcast

Co-authored-by: Michał Pierzchała <thymikee@gmail.com>