fix: doctor for 0.72 Ruby changes
The new release requires support for a range of Ruby version, which are
defined in the project Gemfile. This change add support for validating
against an in project Gemfile (or .ruby-version if it can't find one).
It finally falls back on the baked in version with the CLI.

There are also additional unit tests and some background comments.