feat: check npm and github for template
Support RFC-0759 decoupling of template from react-native. This makes
use of the new @react-native-community/template project. We're careful
to validate against what npm and github show for the state of the
react-native project.

New Template:

This will enable the new @react-native-community/template only if it
published on npm and the react-native/template doesn't exist for the
specific version of react-native selected.

There's more discussion about how template version and react-native
version should be related to each other in #2345.