feat: improve native_module.gradle usage
This change include the ability for project to set their own custom path to the actual folder which contains the react-native project. 

Given we have a consumer app which contains an android project but not actually react-native with the exception of the react-native gradle plugin, the current integration would lead to bunch of build failures. This is because the native_modules.gradle integration assumes that the react native project will be at the same some spot as the gradle root project location.

This change is backwards compatible and won't break any templates or existing projects.