fix: update react-native version in the template for versions >= 0.75. (#2417)
* fix: update react-native version in the template for versions >= 0.75.

The new template handling in @react-native-community/template means that
these are shipped with version 1000.0.0 for React Native. The CLI now
owns this.

* fix: --version & --template handling for 0.75

1. Allow using both of these, but only when the user specifies a version
   of react-native >= 0.75. Since templates are now detached there are
   valid use cases for developers and users to specify a template.
   Added a warning about this test(scope): title
2. Exit early for 0.75+ when we can't find a matching
   @react-native-community/template to the version of react-native.
   These should always be matching, but allow for users to remidy this
   by specifying a template. Clearly this should never happen though.

I've also gone to some effort to use concrete versions in the code when
tags are passed.  E.g. --version next, currently maps to 0.75.0-rc0. We
use the latter when looking up templates.

* fix: update @react-native/ scoped packages to match version

From 0.75 our @react-native/ dependencies must match the version of
react-native. Updated init to perform this on
@react-native-community/template.