refactor: init now expects templates to have react-native version set
For 0.75's previous RCs, we didn't have the infrastructure in place to
tag @react-native-community/templates 0.75-stable branch correctly, as
well as publishing the corresponding version to npm.

For example:

If @react-native-community/template@0.75.1 is the latest version
published, and I specified --version next: I'd expect 0.75.1 of the
template to be used with 0.75.0-rc.1 of react-native.

The downside of this approach is, if you init an older version then
things might not work.  You'd have to specify the matching template
using --template @react-native-community/template@<some version>. Given
that init is intended for new projects this isn't a concern for most
users.

We could look at abusing the `scripts` field to expose the corresponding
version of react native in the npm registry:

  https://registry.npmjs.org/@react-native-community/template

Something like: "scripts": { "version": "0.75.1-rc.2" }